refactor: extract useSongPreviewConfig hook
